refactor: simplify classification.brits;
